requirements.txt
autogen-agentchat>=0.2.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
requests>=2.31.0
python-dotenv>=1.0.0
eventlet>=0.33.0